requirements.txt
Django>=2.0,<3.0
djangorestframework>=3.11
django-heroku>=0.3
dj-database-url
Pillow
psycopg2-binary>=2.0
markdown